Moved deprecated decorator
