+++ sympy/calculus/finite_diff.py
@@ -19,11 +19,11 @@ for:
 
 from sympy.core.function import Derivative
 from sympy.core.singleton import S
-from sympy.core.decorators import deprecated
 from sympy.core.function import Subs
 from sympy.core.traversal import preorder_traversal
 from sympy.utilities.exceptions import SymPyDeprecationWarning
 from sympy.utilities.iterables import iterable
+from sympy.utilities.decorator import deprecated
