+++ sympy/core/symbol.py
@@ -13,7 +13,6 @@ from sympy.logic.boolalg import Boolean
 from sympy.utilities.iterables import sift, is_sequence
 from sympy.utilities.misc import filldedent
 
-
 import string
 import re as _re
 import random
