+++ sympy/integrals/integrals.py
@@ -1,4 +1,3 @@
-from sympy.calculus import singularities
 from sympy.concrete.expr_with_limits import AddWithLimits
 from sympy.core.add import Add
 from sympy.core.basic import Basic
