+++ sympy/physics/matrices.py
@@ -2,7 +2,7 @@
 
 from sympy.core.numbers import I
 from sympy.matrices.dense import MutableDenseMatrix as Matrix
-from sympy.core.decorators import deprecated
+from sympy.utilities.decorator import deprecated
 
 
 def msigma(i):
