+++ sympy/polys/domains/domain.py
@@ -4,7 +4,6 @@
 from typing import Any, Optional, Type
 
 from sympy.core import Basic, sympify
-from sympy.core.decorators import deprecated
 from sympy.core.sorting import default_sort_key, ordered
 from sympy.external.gmpy import HAS_GMPY
 from sympy.polys.domains.domainelement import DomainElement
@@ -12,6 +11,7 @@ from sympy.polys.orderings import lex
 from sympy.polys.polyerrors import UnificationFailed, CoercionFailed, DomainError
 from sympy.polys.polyutils import _unify_gens, _not_a_coeff
 from sympy.utilities import public
+from sympy.utilities.decorator import deprecated
 from sympy.utilities.iterables import is_sequence
