+++ sympy/tensor/tensor.py
@@ -44,7 +44,6 @@ from sympy.combinatorics.tensor_can import get_symmetric_group_sgs, \
 from sympy.core import Basic, Expr, sympify, Add, Mul, S
 from sympy.core.assumptions import ManagedProperties
 from sympy.core.containers import Tuple, Dict
-from sympy.core.decorators import deprecated
 from sympy.core.sorting import default_sort_key
 from sympy.core.symbol import Symbol, symbols
 from sympy.core.sympify import CantSympify, _sympify
@@ -52,7 +51,7 @@ from sympy.core.operations import AssocOp
 from sympy.external.gmpy import SYMPY_INTS
 from sympy.matrices import eye
 from sympy.utilities.exceptions import SymPyDeprecationWarning
-from sympy.utilities.decorator import memoize_property
+from sympy.utilities.decorator import memoize_property, deprecated
 import warnings
