+++ sympy/utilities/decorator.py
@@ -6,12 +6,13 @@ import inspect
 from functools import wraps
 
 from sympy.testing.runtests import DependencyError, SymPyDocTests, PyTestReporter
-from sympy.utilities.iterables import iterable
+from sympy.utilities.exceptions import SymPyDeprecationWarning
 
 def threaded_factory(func, use_add):
     """A factory for ``threaded`` decorators. """
     from sympy.core import sympify
     from sympy.matrices import MatrixBase
+    from sympy.utilities.iterables import iterable
 
     @wraps(func)
     def threaded_func(expr, *args, **kwargs):
@@ -238,3 +239,32 @@ def memoize_property(propfunc):
         return val
 
     return property(accessor)
+
+
+def deprecated(**decorator_kwargs):
+    """This is a decorator which can be used to mark functions
+    as deprecated. It will result in a warning being emitted
+    when the function is used."""
+
+    def _warn_deprecation(wrapped, stacklevel):
+        decorator_kwargs.setdefault('feature', wrapped.__name__)
+        SymPyDeprecationWarning(**decorator_kwargs).warn(stacklevel=stacklevel)
+
+    def deprecated_decorator(wrapped):
+        if hasattr(wrapped, '__mro__'):  # wrapped is actually a class
+            class wrapper(wrapped):
+                __doc__ = wrapped.__doc__
+                __name__ = wrapped.__name__
+                __module__ = wrapped.__module__
+                _sympy_deprecated_func = wrapped
+                def __init__(self, *args, **kwargs):
+                    _warn_deprecation(wrapped, 4)
+                    super().__init__(*args, **kwargs)
+        else:
+            @wraps(wrapped)
+            def wrapper(*args, **kwargs):
+                _warn_deprecation(wrapped, 3)
+                return wrapped(*args, **kwargs)
+            wrapper._sympy_deprecated_func = wrapped
+        return wrapper
+    return deprecated_decorator
